Fix numerical and ID resolvers for lists, reorganise resolvers
