+++ packages/graphql/src/classes/Neo4jGraphQL.ts
@@ -33,7 +33,7 @@ import assertIndexesAndConstraints, {
     AssertIndexesAndConstraintsOptions,
 } from "./utils/asserts-indexes-and-constraints";
 import { wrapResolver, wrapSubscription } from "../schema/resolvers/wrapper";
-import { defaultFieldResolver } from "../schema/resolvers";
+import { defaultFieldResolver } from "../schema/resolvers/field/defaultField";
 import { asArray } from "../utils/utils";
 import { DEBUG_ALL } from "../constants";
