+++ packages/graphql/src/schema/resolvers/mutation/create.test.ts
@@ -17,23 +17,20 @@
  * limitations under the License.
  */
 
-import cypherResolver from "./cypher";
-import { CypherField } from "../../types";
+import { createResolver } from "./create";
+import { NodeBuilder } from "../../../../tests/utils/builders/node-builder";
 
-describe("Cypher resolver", () => {
+describe("Create resolver", () => {
     test("should return the correct; type, args and resolve", () => {
-        // @ts-ignore
-        const field: CypherField = {
-            // @ts-ignore
-            typeMeta: { name: "Test", pretty: "[Test]" },
-            arguments: [],
-            isEnum: false,
-            isScalar: true,
-        };
+        const node = new NodeBuilder({
+            name: "Movie",
+        }).instance();
 
-        const result = cypherResolver({ field, statement: "", type: "Query" });
-        expect(result.type).toEqual(field.typeMeta.pretty);
+        const result = createResolver({ node });
+        expect(result.type).toBe("CreateMoviesMutationResponse!");
         expect(result.resolve).toBeInstanceOf(Function);
-        expect(result.args).toMatchObject({});
+        expect(result.args).toMatchObject({
+            input: "[MovieCreateInput!]!",
+        });
     });
 });
