+++ packages/graphql/src/schema/resolvers/query/aggregate.ts
@@ -18,17 +18,21 @@
  */
 
 import { GraphQLResolveInfo } from "graphql";
-import { execute } from "../../utils";
-import { translateRead } from "../../translate";
-import { Node } from "../../classes";
-import { Context } from "../../types";
-import getNeo4jResolveTree from "../../utils/get-neo4j-resolve-tree";
+import { execute } from "../../../utils";
+import { Node } from "../../../classes";
+import { Context } from "../../../types";
+import { translateAggregate } from "../../../translate";
+import getNeo4jResolveTree from "../../../utils/get-neo4j-resolve-tree";
 
-export default function findResolver({ node }: { node: Node }) {
-    async function resolve(_root: any, args: any, _context: unknown, info: GraphQLResolveInfo) {
+export function aggregateResolver({ node }: { node: Node }) {
+    async function resolve(_root: any, _args: any, _context: unknown, info: GraphQLResolveInfo) {
         const context = _context as Context;
-        context.resolveTree = getNeo4jResolveTree(info, { args });
-        const [cypher, params] = translateRead({ context, node });
+        context.resolveTree = getNeo4jResolveTree(info);
+
+        const [cypher, params] = translateAggregate({
+            context,
+            node,
+        });
 
         const executeResult = await execute({
             cypher,
@@ -37,16 +41,12 @@ export default function findResolver({ node }: { node: Node }) {
             context,
         });
 
-        return executeResult.records.map((x) => x.this);
+        return Object.values(executeResult.records[0] || {})[0];
     }
 
     return {
-        type: `[${node.name}!]!`,
+        type: `${node.aggregateTypeNames.selection}!`,
         resolve,
-        args: {
-            where: `${node.name}Where`,
-            options: `${node.name}Options`,
-            ...(node.fulltextDirective ? { fulltext: `${node.name}Fulltext` } : {}),
-        },
+        args: { where: `${node.name}Where`, ...(node.fulltextDirective ? { fulltext: `${node.name}Fulltext` } : {}) },
     };
 }
